added login, register via firebase

+++ src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {Recipe} from '../recipe.model';
 import {RecipeService} from '../recipe.service';
 import {ActivatedRoute, Params, Router} from '@angular/router';
+import {AuthService} from '../../auth/auth.service';
 
 @Component({
   selector: 'app-recipe-detail',
@@ -14,7 +15,8 @@ export class RecipeDetailComponent implements OnInit {
 
   constructor(private recipeService: RecipeService,
               private route: ActivatedRoute,
-              private router: Router) {
+              private router: Router,
+              private authService: AuthService) {
   }
 
   ngOnInit() {
@@ -31,6 +33,9 @@ export class RecipeDetailComponent implements OnInit {
   }
 
   onDeleteRecipe() {
+    if (!this.authService.isAuthenticated()) {
+      return false;
+    }
     this.recipeService.deleteRecipe(this.id);
     this.router.navigate(['/recipes']);
   }
